Mon, May 13, 2024, 7:27 PM +05:30

+++ readability/readability.c
@@ -14,9 +14,9 @@ int main(void)
     letters = no_of_letters(text);
     words = no_of_words(text);
     sentences = no_of_sentences(text);
-    double L,S,index;
-    L = (letters/words)*100;
-    S = (sentences/words)*100;
+    float L,S,index;
+    L = (letters/(float)words)*100;
+    S = (sentences/(float)words)*100;
     index = 0.0588*L - 0.296*S -15.8;
     if (index>=16)
     {
